Added composition for PlotLineOrBandAxis.
